+++ filefisher/__init__.py
@@ -2,8 +2,8 @@
 
 from importlib.metadata import version as _get_version
 
-from filefinder import _filefinder, _utils, cmip, filters
-from filefinder._filefinder import FileContainer, FileFinder
+from filefisher import _filefinder, _utils, cmip, filters
+from filefisher._filefinder import FileContainer, FileFinder
 
 __all__ = [
     "_filefinder",
@@ -15,7 +15,7 @@ __all__ = [
 ]
 
 try:
-    __version__ = _get_version("filefinder")
+    __version__ = _get_version("filefisher")
     del _get_version
 except Exception:  # pragma: no cover
     # Local copy or not installed with setuptools.
