+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = filefinder
+name = filefisher
 author = Mathias Hauser
 author_email = mathias.hauser@env.ethz.com
 license = MIT
